pyproject.toml
[tool.poetry]
name = "python-ab2-hw-airflow-ilystsov"
version = "0.1.0"
description = ""
authors = ["ilya <lystsov.ilya@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
apache-airflow = "^2.9.1"
radon = "^6.0.1"
pylint = "^3.2.3"
python-dotenv = "^1.0.1"
matplotlib = "^3.9.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
